# openapi.yaml
openapi: "3.0.3"
info:
  title: Notification Relay API
  description: |
    A notification delivery relay that sends alerts through Email, SMS, and Voice
    to on-call groups, with LDAP-based membership syncing and RBAC.

    ## Authentication

    All endpoints require HTTP Basic Auth. Credentials are verified against LDAP.
    The caller's roles are resolved from their LDAP group membership:

    | Role        | Permissions            |
    |-------------|------------------------|
    | `admin`     | admin + publish + read |
    | `publisher` | publish + read         |
    | `reader`    | read only              |
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - basicAuth: []
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
...
  schemas:
...
    SMSSubscription:
      type: object
      required: [username, phone, subscribed_at]
      properties:
        username:
          type: string
          description: The subscribed user's LDAP username
        phone:
          type: string
          description: Mobile phone number in E.164 format
        subscribed_at:
          type: string
          format: date-time
          description: When the user opted in (RFC3339)
      example:
        username: "jsmith"
        phone: "+15551234567"
        subscribed_at: "2026-04-28T10:00:00Z"
...
paths:
...
  /api/v1/sms-subscriptions:
    get:
      summary: List SMS subscriptions
      description: Returns all users who have opted in to SMS notifications. Requires admin role.
      operationId: listSMSSubscriptions
      tags: [SMS Subscriptions]
      responses:
        "200":
          description: All SMS subscribers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SMSSubscription"
        "401":
          description: Authentication required or invalid credentials
        "403":
          description: Forbidden — requires admin role
        "500":
          description: Internal server error

    post:
      summary: Subscribe a user (admin)
      description: |
        Looks up the user's mobile number in LDAP and adds them to the SMS
        subscription list. Returns the existing subscription if the user is
        already subscribed. Also sends the user a welcome SMS via the
        dispatcher. Requires admin role.
      operationId: adminSubscribe
      tags: [SMS Subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  description: LDAP username to subscribe
            example:
              username: "jsmith"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMSSubscription"
        "200":
          description: User was already subscribed; existing record returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMSSubscription"
        "400":
          description: Missing or invalid request body
        "401":
          description: Authentication required or invalid credentials
        "403":
          description: Forbidden — requires admin role
        "404":
          description: User not found in LDAP
        "422":
          description: User has no mobile number on record, or number is not E.164
        "500":
          description: Internal server error
...